+++ src/main/java/ic3/analyticsops/test/AOTaskID.java
@@ -2,6 +2,8 @@ package ic3.analyticsops.test;
 
 public enum AOTaskID
 {
+    ClearAuthDataCache,
+    ClearOnTheFlyPermsCache,
     ClearResultCache,
     DeleteSchemaBackup,
     GenerateMDXes,