+++ src/main/java/ic3/analyticsops/test/task/server/AOClearAuthDataCacheTask.java
@@ -1,0 +1,50 @@
+package ic3.analyticsops.test.task.server;
+
+import ic3.analyticsops.common.AOException;
+import ic3.analyticsops.restapi.request.AORestApiClearAuthDataCacheRequest;
+import ic3.analyticsops.test.AOAssertionMode;
+import ic3.analyticsops.test.AOTask;
+import ic3.analyticsops.test.AOTaskContext;
+import ic3.analyticsops.test.AOTestValidationException;
+
+public class AOClearAuthDataCacheTask extends AOTask
+{
+    protected AOClearAuthDataCacheTask()
+    {
+        // JSON deserialization
+    }
+
+    @Override
+    public void validateProps()
+            throws AOTestValidationException
+    {
+        super.validateProps();
+    }
+
+    @Override
+    public String getKind()
+    {
+        return "ClearAuthDataCache";
+    }
+
+    @Override
+    public AOAssertionMode getAssertionsMode()
+    {
+        return AOAssertionMode.NONE;
+    }
+
+    @Override
+    public void run(AOTaskContext context)
+            throws AOException
+    {
+        context.sendRequest(
+
+                new AORestApiClearAuthDataCacheRequest()
+
+        );
+
+        // Unless an exception has been thrown the cache should have been cleared by now.
+
+    }
+
+}